Documentation

This page lists and describes all arguments of the R package semisup.

Arguments

y
observations: numeric vector of length n

Y
observations: numeric vector of length n, or numeric matrix with n rows (samples) and q columns (variables)

z
class labels: integer vector of length n, with entries 0, 1 and NA

Z
class labels: numeric vector of length n, or numeric matrix with n rows (samples) and p columns (variables), with entries 0 and NA

dist
distributional assumption: character "norm" (Gaussian), "nbinom" (negative bionomial), or "zinb" (zero-inflated negative binomial)

phi
dispersion parameter: positive numeric, or NULL

phi
dispersion parameters: numeric vector of length q, or NULL

pi
zero-inflation parameter: numeric between 0 and 1, or NULL

pi
zero-inflation parameter(s): numeric vector of length q, or NULL

gamma
offset: numeric vector of length n, or NULL

test
resampling procedure: character "perm" (permutation) or "boot" (parametric bootstrap), or NULL

iter
(maximum) number of resampling iterations : positive integer, or NULL

kind
resampling accuracy: numeric between 0 and 1, or NULL; all p-values above kind are approximate

starts
restarts of the EM algorithm: positive integer (defaults to 1)

it.em
(maximum) number of iterations in the EM algorithm: positive integer (defaults to 100)

epsilon
convergence criterion for the EM algorithm: non-negative numeric (defaults to 1e-04)

debug
verification of arguments: TRUE or FALSE

pass
parameters for parametric bootstrap algorithm

...
settings EM algorithm: starts, it.em and epsilon (see arguments)

See also

Use mixtura for model fitting, and scrutor for hypothesis testing. All other functions of the R package semisup are internal.