Documentation

This page lists and describes some internal functions of the R package semisup. These functions should not be used for analysing data.

fit.wrap multiple restarts
fit.norm Gaussian mixture model
fit.nbinom negative binomial mixture model
fit.zinb zero-inflated negative binomial mixture model
estim.nbinom dispersion estimation
estim.zinb dispersion and zero-inflation estimation
resam.lrts resampling (bootstrap, permutation)

See also