Internal function

This function verifies whether the arguments fulfill some formal requirements.

debug(y, z, dist, phi, pi, gamma, test, iter, kind, ...)

Arguments

y
observations: numeric vector of length n

z
class labels: integer vector of length n, with entries 0, 1 and NA

dist
distributional assumption: character "norm" (Gaussian), "nbinom" (negative bionomial), or "zinb" (zero-inflated negative binomial)

phi
dispersion parameters: numeric vector of length q, or NULL

pi
zero-inflation parameter(s): numeric vector of length q, or NULL

gamma
offset: numeric vector of length n, or NULL

test
resampling procedure: character "perm" (permutation) or "boot" (parametric bootstrap), or NULL

iter
(maximum) number of resampling iterations : positive integer, or NULL

kind
resampling accuracy: numeric between 0 and 1, or NULL; all p-values above kind are approximate

...
settings EM algorithm: starts, it.em and epsilon (see arguments)

Value

This function returns warnings and errors. It also returns kind (see details).

Details

If one or more entries of z are equal to 1, the mixture model can be fitted but not tested. Accordingly, kind is replaced by NULL.

Resampling-based testing cannot reach p-values below 1/iter. If kind is smaller than 1/iter, it is replaced by 0.

See also

Examples

NULL

#> NULL